+++ revbrowser/src/cefbrowser_w32.cpp
@@ -521,17 +521,3 @@ bool MCCefWin32Browser::PlatformGetAuthCredentials(bool p_is_proxy, const CefStr
 #include <stdio.h>
 #include <cstring>
 
-void *MCU_loadmodule(const char *p_source)
-{
-    return LoadLibraryA(p_source);
-}
-
-void MCU_unloadmodule(void *p_module)
-{
-    
-}
-
-void *MCU_resolvemodulesymbol(void *p_module, const char *p_name)
-{
-    return GetProcAddress((HMODULE)p_module, p_name);
-}
