+++ revbrowser/src/cefshared_w32.cpp
@@ -1,0 +1,37 @@
+/* Copyright (C) 2015 Runtime Revolution Ltd.
+
+This file is part of LiveCode.
+
+LiveCode is free software; you can redistribute it and/or modify it under
+the terms of the GNU General Public License v3 as published by the Free
+Software Foundation.
+
+LiveCode is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or
+FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+for more details.
+
+You should have received a copy of the GNU General Public License
+along with LiveCode.  If not see <http://www.gnu.org/licenses/>.  */
+
+#include <windows.h>
+
+extern "C"
+{
+
+void *MCU_loadmodule(const char *p_source)
+{
+    return LoadLibraryA(p_source);
+}
+
+void MCU_unloadmodule(void *p_module)
+{
+    
+}
+
+void *MCU_resolvemodulesymbol(void *p_module, const char *p_name)
+{
+    return GetProcAddress((HMODULE)p_module, p_name);
+}
+    
+}
